fix(client-timeout): change default timeout to no timeout and make sure go-api respects it
This has been discussed (Slack link below) and seems like there really is no reason to have any sort of a timeout in the client. Our backend and load balancers already enforce timeouts on the backend side and there seems to be no benefit in timing out a client request in the client end, as that will only throw errors to the users that don't really convey any real information. In particular, storage imports can easily go over any sensible timeouts so the other option would be to have per-command timeouts.

Ref: https://upcloud.slack.com/archives/C01MSUTR6PL/p1617269550033700