fix(import/windows): fix parsing of source-location on Windows
paths like 'c:\foo\bar' seem to be parsed as urls (with scheme 'c' and the rest of the path going
to opaque field), which causes them to fail with a weird message. Change the logic to try to stat()
the given source-location first, and if found, consider it a local upload.