fix: snyk errors and missing mock services (#540)
The build check were failing because of 3rd party dependency coming soon
from upctl deploy starter-kit.
Changes in the sdk made the mock service used in the testing to fail
because some classes where missing.
list_test.go, part of the server tests was failing because of changes
made for the floating ip address in the sdk.
I manually upgraded to latest sdk release